chore: more test alterations based on consistency
